fix: retry when unable to find content from endpoints
fixes #751

the original `maxRetries` in the `getJSONWatchPage` function was only
for that endpoint, did not wait for the next request, and was not
customizable. now, retries all included for the main 3 endpoints
- watch.json
- embed page
- get_video_info
these retries are customized through miniget's similar `maxRetries` and
`backoff` options. these shared options were used since these failing
scenarios are also failing requests.