fix: fix race condition causing multiple watch page requests
this was happening due to the `bpctr` param added to bypass the
"offensive video" warning. if the `getInfo()` process took longer than a
second, it would make another request.

the `bpctr` param is now removed since it no longer seems to be needed.
if you find a video that still shows a warning, let me know.